+++ usr/src/sys/net/if_slvar.h
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)if_slvar.h	8.3 (Berkeley) %G%
+ *	@(#)if_slvar.h	8.4 (Berkeley) %G%
  *
  * $Header: if_slvar.h,v 1.3 89/05/31 02:25:18 van Exp $
  */
@@ -45,10 +45,10 @@ struct sl_softc {
 void	slattach __P((void));
 void	slclose __P((struct tty *));
 void	slinput __P((int, struct tty *));
-int	slioctl __P((struct ifnet *, int, caddr_t));
+int	slioctl __P((struct ifnet *, u_long, caddr_t));
 int	slopen __P((dev_t, struct tty *));
 int	sloutput __P((struct ifnet *,
 	    struct mbuf *, struct sockaddr *, struct rtentry *));
 void	slstart __P((struct tty *));
-int	sltioctl __P((struct tty *, int, caddr_t, int));
+int	sltioctl __P((struct tty *, u_long, caddr_t, int));
 #endif /* KERNEL */
